+++ src/main/java/com/ssginc8/docto/user/service/UserServiceImpl.java
@@ -19,14 +19,14 @@ import com.ssginc8.docto.file.entity.File;
 import com.ssginc8.docto.file.service.FileService;
 import com.ssginc8.docto.global.error.exception.userException.DuplicateEmailException;
 import com.ssginc8.docto.global.error.exception.userException.InvalidPasswordException;
-import com.ssginc8.docto.user.dto.AddDoctorList;
-import com.ssginc8.docto.user.dto.AddUser;
-import com.ssginc8.docto.user.dto.FindEmail;
-import com.ssginc8.docto.user.dto.Login;
-import com.ssginc8.docto.user.dto.SocialSignup;
 import com.ssginc8.docto.user.entity.Role;
 import com.ssginc8.docto.user.entity.User;
 import com.ssginc8.docto.user.provider.UserProvider;
+import com.ssginc8.docto.user.service.dto.AddDoctorList;
+import com.ssginc8.docto.user.service.dto.AddUser;
+import com.ssginc8.docto.user.service.dto.FindEmail;
+import com.ssginc8.docto.user.service.dto.Login;
+import com.ssginc8.docto.user.service.dto.SocialSignup;
 import com.ssginc8.docto.user.validator.CreateUserValidator;
 
 import lombok.RequiredArgsConstructor;
