+++ src/main/java/com/ssginc8/docto/user/service/dto/FindEmail.java
@@ -1,4 +1,4 @@
-package com.ssginc8.docto.user.dto;
+package com.ssginc8.docto.user.service.dto;
 
 import lombok.Builder;
 import lombok.Getter;
